Refactor performance report form
