+++ src/data/RealData.tsx
@@ -158,7 +158,7 @@ export const allLogs: Array<Log> = [
     taRatedEmail: "alex.ference@mail.mcgill.ca",
     courseID: "5f13eef5-b289-4a15-8e7d-b0dfd3715358",
     rating: 4,
-    comment: "Was very helpful conducting zoom class for me.",
+    comments: ["Was very helpful conducting zoom class for me.", "Covered my office hours this week - thanks!"],
     createdAt: "2022-04-19 16:02:04.547 +00:00",
     updatedAt: "2022-04-19 16:02:04.547 +00:00",
   },
@@ -167,7 +167,7 @@ export const allLogs: Array<Log> = [
     taRatedEmail: "ddorantesferreira@gmail.com",
     courseID: "1ec6382a-8998-499f-b325-df9ec109a9f5",
     rating: 3,
-    comment: "Covered my office hours this week - thanks!",
+    comments: ["Was very helpful conducting zoom class for me.", "Covered my office hours this week - thanks!"],
     createdAt: "2022-04-19 17:47:08.716 +00:00",
     updatedAt: "2022-04-19 17:47:08.716 +00:00",
   },
